Replace domainparticipantasync listener by dyn version
