+++ dds/src/implementation/actors/domain_participant_actor.rs
@@ -5,8 +5,9 @@ use crate::{
     builtin_topics::{BuiltInTopicKey, ParticipantBuiltinTopicData, TopicBuiltinTopicData},
     dds::infrastructure,
     dds_async::{
-        domain_participant::DomainParticipantAsync, publisher_listener::PublisherListenerAsync,
-        subscriber_listener::SubscriberListenerAsync,
+        domain_participant::DomainParticipantAsync,
+        domain_participant_listener::DomainParticipantListenerAsync,
+        publisher_listener::PublisherListenerAsync, subscriber_listener::SubscriberListenerAsync,
     },
     domain::domain_participant_factory::DomainId,
     implementation::{
@@ -83,9 +84,7 @@ use std::{
 use super::{
     data_reader_actor,
     data_writer_actor::{self, DataWriterActor},
-    domain_participant_listener_actor::{
-        DomainParticipantListenerActor, DomainParticipantListenerAsyncDyn,
-    },
+    domain_participant_listener_actor::DomainParticipantListenerActor,
     publisher_actor::{self, PublisherActor},
     status_condition_actor::StatusConditionActor,
     subscriber_actor, topic_actor,
@@ -250,7 +249,7 @@ impl DomainParticipantActor {
         spdp_discovery_locator_list: &[Locator],
         data_max_size_serialized: usize,
         udp_transport_write: Arc<UdpTransportWrite>,
-        listener: Box<dyn DomainParticipantListenerAsyncDyn + Send>,
+        listener: Box<dyn DomainParticipantListenerAsync + Send>,
         status_kind: Vec<StatusKind>,
         handle: &tokio::runtime::Handle,
     ) -> Self {
@@ -1367,7 +1366,7 @@ impl DomainParticipantActor {
 
     async fn set_listener(
         &mut self,
-        listener: Box<dyn DomainParticipantListenerAsyncDyn + Send>,
+        listener: Box<dyn DomainParticipantListenerAsync + Send>,
         status_kind: Vec<StatusKind>,
         runtime_handle: tokio::runtime::Handle,
     ) {
